requirements.txt
requests
lxml
pymongo
pandas
python-dotenv
aiokafka